feat: add exports of deleteUser thunk, and setup extraReducer
deleteUser thunk was created in previous commit, now
the exports of deleteUser will be added from store/index.js
by this commit. This commit will also setup the extraReducers
for deleteUser.

The extraReducer builder.addCase function is not ready yet,
and needs to work on. But before we need to do some refactor
to the UsersList component. The extraReducer function will be
built in upcoming commits.