feat: create delete user mechanism
As thunk func for deleting a user was already created, with a small
bug, that it returns the responce from axios.delete, which is pointless,
as we already have access to the user object intended to be deleted as a prop
, so this commit fixes that to be returning only the user object from the Thunk.
Then in usersSlice, the extraReducer initiated for deleteUser.fulfilled will have
a updated array by using the filter method on the state.data, filtering every
object that is not the intended object to delete. And then updating the state.data
with the updated array. Then in the UsersListItem component this commit creates, a
decent UX for delete button and uses useThunkOperator hook to dispatch deleteUser
and keep loading screens. This component calls the deleteUser with the user as an
argument through useThunkOperator hook.