feat: create component to handle expandable box in user list
Our app needs the user list items to be expanded so that they can contain,
albums list in them, and a heading and a button. In order to achieve that
we need another component for handling exandable panel's logics. This commit
creates a component with the exact reason and calls in UsersListItem.