fix: specify user in endpoints tag to prevent multiple req sendings
The provdesTags argument in endpoints are recieving and sending only
a string, which makes the application less efficient as when multiple
users are expanded and one add album is clickes, all the rendered album
lists gets fetched. To prevent this, this commit will pass an object
in providesTags and invalidatesTags arguments with a type(string) and
an id to specify which user's data is changing. And to follow that change
to make a new fetch request.