feat: add endpoint for handling image creation
This commit writes all the neccessary code for creating an endpoint
for handling image creations. Then that endpoint hook func is also
used in ImageList component's add image's invoked func to call the
endpoint func. This commit also adds the providesTags and invalidatesTags
in the fetchImages and createImages endpoints respectively. Allover this
commit creates all neccessary logics for image creation and rendering
feature.