feat: create image deletation endpoint
This commit will create another endpoint for handling deletations of
images. This commit also imports that endpoint hook function in ImageListItem
component and calls that in the Delete Button's invoked function. This commit
will also take care of the tags, so that after invoking the deletation function,
the component re-renders.