fix: add userId to image objs to avoid wrong associations
Currently the image objects have a albumId that specifies which album
that image belongs to. And the general theory is that every album also
has a userId to specify which user they belong to. And everything should
render properly on this hiararchy. But when deleting a user, the albums
inside it gets deleted, but not the images. As a side effect to that, if
another album get's the id of the deleted album, then the images(which
should have been deleted) renders under this new assigned album.

The bug in short is, when deleting a user, it should delete all albums and
images inside it. But it seems like, that is not the case, and the images
are not getting deleted.

To fix that problem this commit will create another prop in image objects
as userId to specify, which user that image belong to. That prop will be
derived directly from album object's userId. This commit will also add
the userId prop in fetchImages endpoint query function. And that solves
the issue.

Issue #1 Solved